+++ examples/notifications/pages/_app.tsx
@@ -1,6 +1,6 @@
 import 'tailwindcss/tailwind.css';
-import '../styles/globals.css';
 import '@dialectlabs/react-ui/index.css';
+import '../styles/globals.css';
 import React from 'react';
 import type { AppProps } from 'next/app';