+++ packages/dialect-react-ui/utils/useMobile.ts
@@ -1,0 +1,5 @@
+import { useMediaPredicate } from 'react-media-hook';
+
+export default function useMobile() {
+  return useMediaPredicate('(max-width: 640px)');
+}